feat: do hybrid ts / json
