+++ packages/language-server/package.json
@@ -10,12 +10,7 @@
   },
   "type": "commonjs",
   "main": "dist/index.js",
-  "files": [
-    "bin",
-    "dist/**/*.js",
-    "dist/**/*.d.ts",
-    "types/**/*.d.ts"
-  ],
+  "files": ["bin", "dist/**/*.js", "dist/**/*.d.ts", "types/**/*.d.ts"],
   "bin": {
     "astro-ls": "./bin/nodeServer.js"
   },
@@ -27,13 +22,13 @@
   },
   "dependencies": {
     "@astrojs/compiler": "^2.10.2",
+    "@astrojs/yaml2ts": "^0.1.0",
     "@jridgewell/sourcemap-codec": "^1.4.15",
     "@volar/kit": "~2.4.0-alpha.15",
     "@volar/language-core": "~2.4.0-alpha.15",
     "@volar/language-server": "~2.4.0-alpha.15",
     "@volar/language-service": "~2.4.0-alpha.15",
     "@volar/typescript": "~2.4.0-alpha.15",
-    "devalue": "^5.0.0",
     "fast-glob": "^3.2.12",
     "muggle-string": "^0.4.1",
     "volar-service-css": "0.0.59",
@@ -42,9 +37,9 @@
     "volar-service-prettier": "0.0.59",
     "volar-service-typescript": "0.0.59",
     "volar-service-typescript-twoslash-queries": "0.0.59",
+    "volar-service-yaml": "0.0.59",
     "vscode-html-languageservice": "^5.2.0",
-    "vscode-uri": "^3.0.8",
-    "yaml": "^2.4.5"
+    "vscode-uri": "^3.0.8"
   },
   "devDependencies": {
     "@astrojs/svelte": "^5.0.3",
