+++ packages/language-server/src/core/utils.ts
@@ -2,6 +2,8 @@ import type { VirtualCode } from '@volar/language-core';
 import { URI, Utils } from 'vscode-uri';
 import { importSvelteIntegration, importVueIntegration } from '../importPackage';
 
+export const FRONTMATTER_OFFSET = 4;
+
 export function framework2tsx(
 	filePath: string,
 	sourceCode: string,
