+++ packages/language-server/src/languageServerPlugin.ts
@@ -5,7 +5,7 @@ import {
 	MessageType,
 	ShowMessageNotification,
 } from '@volar/language-server/node';
-import { URI } from 'vscode-uri';
+import { URI, Utils } from 'vscode-uri';
 import { getAstroLanguagePlugin } from './core';
 import { getSvelteLanguagePlugin } from './core/svelte.js';
 import { getVueLanguagePlugin } from './core/vue.js';
@@ -18,12 +18,13 @@ import { create as createEmmetService } from 'volar-service-emmet';
 import { create as createPrettierService } from 'volar-service-prettier';
 import { create as createTypeScriptTwoSlashService } from 'volar-service-typescript-twoslash-queries';
 
-import { getFrontmatterLanguagePlugin } from './frontmatter/languagePlugin.js';
+import { getFrontmatterLanguagePlugin } from './core/frontmatterHolders.js';
 import { create as createAstroService } from './plugins/astro.js';
 import { create as createHtmlService } from './plugins/html.js';
 import { create as createMarkdownService } from './plugins/markdown.js';
 import { create as createTypescriptAddonsService } from './plugins/typescript-addons/index.js';
 import { create as createTypeScriptServices } from './plugins/typescript/index.js';
+import { create as createYAMLService } from './plugins/yaml.js';
 
 export function getLanguagePlugins(
 	connection: Connection,
@@ -55,7 +56,7 @@ export function getLanguagePlugins(
 
 	languagePlugins.unshift(
 		getAstroLanguagePlugin(typeof astroInstall === 'string' ? undefined : astroInstall, ts),
-		getFrontmatterLanguagePlugin(typeof astroInstall === 'string' ? undefined : astroInstall, ts)
+		getFrontmatterLanguagePlugin()
 	);
 
 	return languagePlugins;
@@ -72,6 +73,7 @@ export function getLanguageServicePlugins(connection: Connection, ts: typeof imp
 		createAstroService(ts),
 		getPrettierService(),
 		createMarkdownService(),
+		createYAMLService(),
 	];
 	function getPrettierService() {
 		let prettier: ReturnType<typeof importPrettier>;
