+++ packages/yaml2ts/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "@astrojs/yaml2ts",
+  "version": "0.1.0",
+  "description": "Package to convert a YAML string to a Volar VirtualCode, to be used by the Astro language server to provide intellisense",
+  "scripts": {
+    "build": "tsc",
+    "dev": "tsc --watch"
+  },
+  "type": "commonjs",
+  "main": "dist/index.js",
+  "files": ["dist/**/*.js", "dist/**/*.d.ts"],
+  "keywords": [],
+  "author": "",
+  "license": "MIT",
+  "dependencies": {
+    "yaml": "^2.5.0",
+    "devalue": "^5.0.0"
+  },
+  "devDependencies": {
+    "@volar/language-core": "~2.4.0-alpha.15",
+    "typescript": "^5.2.2"
+  }
+}
