+++ packages/yaml2ts/src/index.ts
@@ -1,0 +1,1 @@
+export { yaml2ts } from './yaml2ts';
